correct decay state dtype to int

+++ dental_env/envs/dental_env_5d.py
@@ -56,7 +56,7 @@ class DentalEnv5D(gym.Env):
         self._agent_location = np.append(self.np_random.integers(0, self.size, size=2),
                                          [self.size - 1, 0, 0]).astype(int)  # start from random
         # state initialization
-        self._states = np.ones((self.size, self.size, self.size)) * 2
+        self._states = np.ones((self.size, self.size, self.size), dtype=int) * 2
         decay_position = self.np_random.integers(low=[1, 1, 0], high=[self.size - 1, self.size - 1, self.size - 1],
                                                  size=(5, 3))
         decay_size = self.np_random.integers(low=[1, 1, 1],
